% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_adcfit.R
\name{camino_adcfit}
\alias{camino_adcfit}
\title{Wrapper for Camino \code{adcfit} function}
\usage{
camino_adcfit(infile, scheme, outfile = NULL, outputdatatype = c("double",
  "float", "byte", "short", "int", "long"), verbose = TRUE)
}
\arguments{
\item{infile}{input filename}

\item{scheme}{scheme file for b-values/b-vectors}

\item{outfile}{Output filename}

\item{outputdatatype}{Output data type}

\item{verbose}{Print diagnostic messages}
}
\description{
Performs the Camino \code{adcfit} function
}
