% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_cmd.R
\name{camino_cmd}
\alias{camino_cmd}
\title{General Camino Command Maker}
\usage{
camino_cmd(funcname)
}
\arguments{
\item{funcname}{Name of function to be called}
}
\value{
Character vector of length 1
}
\description{
Wrapper for finding the path for the Camino function
}
\examples{
camino_cmd("voxel2scanner")
}

