% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_compute.R
\name{camino_compute}
\alias{camino_compute}
\title{Wrapper for Computing FA/MD and other Standard Markers}
\usage{
camino_compute(cmd = c("fa", "md"), infile, inputmodel = c("dt",
  "twotensor", "threetensor", "multitensor"), outfile = NULL,
  inputdatatype = NULL, outputdatatype = c("double", "float", "char",
  "short", "int", "long"), verbose = TRUE)
}
\arguments{
\item{cmd}{Command to perform}

\item{infile}{Output from \code{\link{camino_modelfit}}}

\item{inputmodel}{model used for tensors}

\item{outfile}{Output filename for FA map}

\item{inputdatatype}{specify input data type, options are
float, char, short, int, long, double}

\item{outputdatatype}{output data type}

\item{verbose}{print diagnostic messages}
}
\description{
Performs the Camino \code{fa} and \code{md} functions to
compute mean diffusivity (MD) and fractional anisotropy (FA) markers
}

