% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_dt2nii.R
\name{camino_dt2nii}
\alias{camino_dt2nii}
\title{Wrapper for Camino \code{dt2nii} function}
\usage{
camino_dt2nii(infile, header, inputmodel = c("dt", "twotensor",
  "threetensor", "multitensor"), outputroot = NULL,
  outputdatatype = NULL, maxcomponents = NULL, gzip = TRUE,
  verbose = TRUE)
}
\arguments{
\item{infile}{Output from \code{\link{camino_modelfit}}}

\item{header}{NIfTI image with header information}

\item{inputmodel}{model used for tensors}

\item{outputroot}{Output file root. If there is one component, or output is to a vector,
the output image is called root.ext, otherwise output is numbered by component.}

\item{outputdatatype}{Specifies the data type of the output image.
The data type can be any of the following strings:
"byte", "short", "int", "long", "float" or "double".}

\item{maxcomponents}{Maximum tensor components, for use with input model "multitensor".}

\item{gzip}{Set gzip compression for the output image, if supported by the image format.}

\item{verbose}{print diagnostic messages}
}
\description{
Performs the Camino \code{dt2nii} function
}
