% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_dteig.R
\name{camino_dteig}
\alias{camino_dteig}
\title{Wrapper for Camino \code{dteig} function}
\usage{
camino_dteig(infile, outfile = NULL, inputmodel = c("dt", "twotensor",
  "threetensor", "multitensor"), outputdatatype = c("float", "byte", "short",
  "int", "long", "double"), maxcomponents = NULL, verbose = TRUE)
}
\arguments{
\item{infile}{Output from \code{\link{camino_modelfit}}}

\item{outfile}{Output filename.  If \code{NULL}, a temporary filename will be make}

\item{inputmodel}{model used for tensors}

\item{outputdatatype}{Specifies the data type of the output image.
The data type can be any of the following strings:
"byte", "short", "int", "long", "float" or "double".}

\item{maxcomponents}{Maximum tensor components, for use with input model "multitensor".}

\item{verbose}{print diagnostic messages}
}
\description{
Performs the Camino \code{dteig} function
}
