% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_dtifit.R
\name{camino_dtifit}
\alias{camino_dtifit}
\title{Fit Model using Camino}
\usage{
camino_dtifit(infile, bvecs, bvals, mask, model = c("dt", "ldt", "nldt_pos",
  "nldt", "ldt_wtd", "restore"), verbose = TRUE, inputmodel = c("dt",
  "twotensor", "threetensor", "multitensor"), ...)
}
\arguments{
\item{infile}{Input 4D image}

\item{bvecs}{The file containing the b-vectors.}

\item{bvals}{The file containing the b-values.}

\item{mask}{Provides the name of a file containing a brain / background mask.
The file can be raw binary or a NIFTI image.
Raw binary files must be big endian; the default data type is 16-bit shorts,
 but can be changed using the -maskdatatype option.
 The program does not process background voxels, but outputs the same
 number of values in background voxels and foreground voxels.
 Each value is zero in background voxels apart from the exit code which is -1.}

\item{model}{Type of model used: 1-tensor: dt (linear diffusion tensor, default),
ldt - same as dt
nldt_pos (nonlinear optimization, constrained to be positive semi-definite),
nldt     (unconstrained nonlinear optimization),
ldt_wtd  (weighted linear)}

\item{verbose}{Print diagnostic messages}

\item{inputmodel}{model used for tensors, passed to
\code{\link{camino_compute}}}

\item{...}{Additional arguments passed to \code{\link{camino_fsl2scheme}}
or \code{\link{camino_modelfit}} or \code{\link{camino_dt2nii}}}
}
\value{
List of values and such
}
\description{
Fit Model using Camino
}
\examples{
\dontrun{
library(neurobase)
library(rcamino)
img = "~/Downloads/data.nii.gz"
bvals = "~/Downloads/bvals"
bvecs = "~/Downloads/bvecs"
mask = "~/Downloads/nodif_brain_mask.nii.gz"
gradadj = "~/Downloads/grad_dev.nii.gz"
sub = subset_dti(img = img, bvals = bvals, bvecs = bvecs,
                 maximum = 1500,
b_step = 1)
infile = sub$img
bvals = sub$bvals
bvecs = sub$bvecs
inputmodel = "dt"
verbose = TRUE
args = list()
camino_set_heap(heap_size = 10000)
res = camino_dtifit(infile = infile, bvecs = bvecs,
bvals = bvals, mask = mask, model = "dt", inputmodel = "dt")
}
}
