% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_compute.R
\name{camino_fa}
\alias{camino_fa}
\title{Compute Fractional Anisotropy (FA) Maps}
\usage{
camino_fa(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{camino_compute}}}
}
\description{
Performs the Camino \code{fa} function to compute
Fractional Anisotropy maps
}

