% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_fsl2scheme.R
\name{camino_fsl2scheme}
\alias{camino_fsl2scheme}
\title{Wrapper for Camino \code{fsl2scheme} function}
\usage{
camino_fsl2scheme(bvecs, bvals, outfile = NULL, diffusiontime = NULL,
  bscale = NULL, flipx = FALSE, flipy = FALSE, flipz = FALSE,
  interleave = FALSE, zerobval = NULL, numscans = NULL,
  usegradmod = FALSE, verbose = TRUE)
}
\arguments{
\item{bvecs}{The file containing the b-vectors.}

\item{bvals}{The file containing the b-values.}

\item{outfile}{output filename for fsl scheme file}

\item{diffusiontime}{Diffusion time in seconds}

\item{bscale}{Scaling  factor  to  convert  the b-values into different units.
Default is 10^6.}

\item{flipx}{Negate the x component of all the vectors.}

\item{flipy}{Negate the y component of all the vectors.}

\item{flipz}{Negate the z component of all the vectors.}

\item{interleave}{Interleave repeated scans.
Only used with \code{-numscans}. If this  is
selected,  the output will be interleaved, so you will have mea-
surement 0 repeated numScans times, then measurement 1, etc.}

\item{zerobval}{Set an effective zero b-value, input b-values less than or equal
to  this  are set to zero in the output. This is needed for some
Camino programs that normalize the input data by  dividing  each
measurement by the mean b=0 measurement. This  value is tested after any applicable scaling is applied to
the bvals.}

\item{numscans}{Output all measurements number times, used when combining multiple  scans  from the same imaging session. The default behaviour
is to repeat the entire block of measurements,  like  you'd  get
from copying and pasting the scheme number times. If -interleave
is specified, then identical measurements are grouped  together.}

\item{usegradmod}{Use the gradient magnitude to scale b.
This option has no effect if your gradient directions have unit magnitude.
It should  only be  used  if your scanner does not normalize the
gradient directions.}

\item{verbose}{Print diagnostic messages}
}
\value{
Output filename
}
\description{
Performs the Camino \code{fsl2scheme} function
}
