% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_set_heap.R
\name{camino_get_heap}
\alias{camino_get_heap}
\title{Get Camino Heap Size}
\usage{
camino_get_heap(heap_size = 4000)
}
\arguments{
\item{heap_size}{Heap size (in MB) if \code{CAMINO_HEAP_SIZE} is not set}
}
\value{
Numeric Value
}
\description{
Get Camino Heap Size
}
\examples{
camino_get_heap()
}

