% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_help.R
\name{camino_help}
\alias{camino_help}
\title{General Camino Command Help}
\usage{
camino_help(funcname, ...)
}
\arguments{
\item{funcname}{Name of function to be called}

\item{...}{options to pass to \code{\link{system}}}
}
\value{
Result from system
}
\description{
Wrapper for finding the manual for the Camino function
}
\examples{
camino_help("voxel2scanner")
}
