% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_compute.R
\name{camino_md}
\alias{camino_md}
\title{Compute Mean Diffusivity (MD) Maps}
\usage{
camino_md(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{camino_compute}}}
}
\description{
Performs the Camino \code{md} function to compute
Mean Diffusivity maps
}

