% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_pointset2scheme.R
\name{camino_pointset2scheme}
\alias{camino_pointset2scheme}
\title{Wrapper for Camino \code{pointset2scheme} function}
\usage{
camino_pointset2scheme(infile, bvalue = 1e+09, outfile = NULL,
  verbose = TRUE)
}
\arguments{
\item{infile}{Input gradient file}

\item{bvalue}{B-value multiplier}

\item{outfile}{Output filename for scheme file}

\item{verbose}{print diagnostic messages}
}
\value{
Output filename
}
\description{
Performs the Camino \code{pointset2scheme} function
}

