% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_selectshells.R
\name{camino_selectshells}
\alias{camino_selectshells}
\title{Wrapper for Camino \code{selectshells} function}
\usage{
camino_selectshells(infile, schemefile, max_bval = NULL,
  min_bval = NULL, output_root = tempfile(),
  max_unweighted_bval = NULL, remove_zero = FALSE, verbose = TRUE)
}
\arguments{
\item{infile}{Input 4D file to select shells}

\item{schemefile}{Scheme file with gradient information}

\item{max_bval}{Maximum b-value to include in the output. Default infinity
(include all b-values). In the same units as the scheme file.}

\item{min_bval}{Minimum b-value to include in the output.
Default 0. In the same units as the scheme file.
b=0 measurements are treated as a special case and are not removed unless
the \code{remove_zero} is \code{TRUE}.}

\item{output_root}{Root of output files, selected shells of
DWI data are output along with the associated scheme file.
Output file format will be the same as the input (NIFTI or raw).}

\item{max_unweighted_bval}{Maximum b-value to include as an unweighted
measurement. Some imaging schemes have a small non-zero b-value
for the unweighted measurements. With this option, any b-value
less than or equal to the specified value are treated as b=0.}

\item{remove_zero}{If \code{TRUE}, b=0 measurements are not selected.
By default they are selected even if a minimum b-value is specified.}

\item{verbose}{Print diagnostic messages}
}
\value{
Output filename
}
\description{
Performs the Camino \code{selectshells} function
}
