% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_set_heap.R
\name{camino_set_heap}
\alias{camino_set_heap}
\title{Set Camino Heap Size}
\usage{
camino_set_heap(heap_size = 4000)
}
\arguments{
\item{heap_size}{Heap size (in MB) for camino command}
}
\description{
Set Camino Heap Size to Increase Memory
}
\examples{
camino_set_heap(heap_size = 4000)
}
