% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_split4dnii.R
\name{camino_split4dnii}
\alias{camino_split4dnii}
\title{Wrapper for Camino \code{split4dnii} split4dnii}
\usage{
camino_split4dnii(infile, output_root = tempfile(fileext = "_"),
  schemefile = NULL, verbose = TRUE)
}
\arguments{
\item{infile}{Input 4D file to split}

\item{output_root}{Root of output files
Output file format will be the same as the input (NIFTI or raw).}

\item{schemefile}{Scheme file with gradient information (optional).
Will output CSV not returned by this function}

\item{verbose}{Print diagnostic messages}
}
\value{
Output filenames
}
\description{
Performs the Camino \code{selectshells} function
}
