% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_subset_max_bval.R
\name{camino_subset_max_bval}
\alias{camino_subset_max_bval}
\title{Subset Maximum B-value}
\usage{
camino_subset_max_bval(infile, schemefile, max_bval = NULL, verbose = TRUE)
}
\arguments{
\item{infile}{Input 4D image}

\item{schemefile}{Scheme file with b-values and b-vectors}

\item{max_bval}{Maximum b-value to include in the output. Default infinity
(include all b-values). In the same units as the scheme file.}

\item{verbose}{Print diagnostic messages}
}
\value{
List of output filename and output scheme file
}
\description{
Subset Maximum B-value
}
