% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_compute.R
\name{camino_trd}
\alias{camino_trd}
\title{Compute Tensor Trace Maps}
\usage{
camino_trd(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{camino_compute}}}
}
\description{
Performs the Camino \code{trd} function to compute
Tensor Trace maps
}
