% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_compute_img.R
\name{camino_trd_img}
\alias{camino_trd_img}
\title{Compute Tensor Trace Maps}
\usage{
camino_trd_img(infile, header, ...)
}
\arguments{
\item{infile}{Output from \code{\link{camino_modelfit}}}

\item{header}{NIfTI image for header information to pass to
\code{\link{camino_voxel2image}}}

\item{...}{additional arguments to pass to \code{\link{camino_compute}}}
}
\description{
Performs the Camino \code{trd} function to compute
Tensor Trace maps
}

