% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_voxel2image.R
\name{camino_voxel2image}
\alias{camino_voxel2image}
\title{Wrapper for Camino \code{voxel2image} function}
\usage{
camino_voxel2image(infile, header, outputroot = NULL,
  inputdatatype = NULL, outputdatatype = NULL, components = 1,
  gzip = TRUE, outputvector = FALSE, verbose = TRUE)
}
\arguments{
\item{infile}{Input file (double, float, etc.)}

\item{header}{NIfTI image with header information}

\item{outputroot}{Output file root. If there is one component, or output is to a vector,
the output image is called root.ext, otherwise output is numbered by component.}

\item{inputdatatype}{specify input data type, options are
"byte", "short", "int", "long", "float" or "double"}

\item{outputdatatype}{Specifies the data type of the output image.
The data type can be any of the following strings:
"byte", "short", "int", "long", "float" or "double".}

\item{components}{The number of components in the input data. The default is 1.}

\item{gzip}{Set gzip compression for the output image, if supported by the image format.}

\item{outputvector}{Output multi-component voxels to a single image.}

\item{verbose}{print diagnostic messages}
}
\description{
Performs the Camino \code{voxel2image} function
}
