% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_voxelclassify.R
\name{camino_voxelclassify}
\alias{camino_voxelclassify}
\title{Wrapper for Camino \code{voxelclassify} function}
\usage{
camino_voxelclassify(infile, scheme, outfile = NULL, outputdatatype = NULL,
  mask = NULL, bgthresh = NULL, csfthresh = NULL, order = 4,
  ftest = NULL, verbose = TRUE)
}
\arguments{
\item{infile}{input filename}

\item{scheme}{Scheme file}

\item{outfile}{output filename}

\item{outputdatatype}{Specifies the data type of the output image, only
relevant if \code{outfile} is not specified
The data type can be any of the following strings:
"byte", "short", "int", "long", "float" or "double".}

\item{mask}{Character or \code{nifti} to mask out background.}

\item{bgthresh}{background threshold value.  If \code{mask} is specified,
this is overridden.  Ignored  if \code{ftest} is not specified.}

\item{csfthresh}{CSF threshold value}

\item{order}{Set the maximum even spherical harmonic order
that will be considered by the program.
Must be a positive even number. Default is 4.}

\item{ftest}{Specifies  the  F-test thresholds for adopting higher order models. A threshold must be provided for every order between 0  and
the maximum order. The program will not consider a higher order
model unless the result of the F-test is smaller than the
relevant threshold.}

\item{verbose}{Print diagnostic messages}
}
\description{
Performs the Camino \code{voxelclassify} function
}

