% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_opts.R
\name{collapse_log_opts}
\alias{collapse_log_opts}
\title{Collapse Logical Options}
\usage{
collapse_log_opts(opts)
}
\arguments{
\item{opts}{Options to be passed, where the names of the options are the
Camino options}
}
\value{
String
}
\description{
Simple wrapper to collapse options into a string for Camino
}
