% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinetwofibreluts.R
\name{combinetwofibreluts}
\alias{combinetwofibreluts}
\title{Wrapper for Camino \code{combinetwofibreluts} function}
\usage{
combinetwofibreluts()
}
\description{
Performs the Camino \code{combinetwofibreluts} function
}

