% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_camino_args.R
\name{parse_camino_args}
\alias{parse_camino_args}
\title{Parse arguments from Camino}
\usage{
parse_camino_args(funcname)
}
\arguments{
\item{funcname}{Name of function to be called}
}
\value{
Character vector
}
\description{
Wrapper for finding the arguments for the Camino function
}
\examples{
read_camino_helpfile("voxel2scanner")
}

