% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointset2scheme.R
\name{pointset2scheme}
\alias{pointset2scheme}
\title{Wrapper for Camino \code{pointset2scheme} function}
\usage{
pointset2scheme(infile, bvalue = NULL, outfile = NULL)
}
\description{
Performs the Camino \code{pointset2scheme} function
}

