% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_help.R
\name{read_camino_helpfile}
\alias{read_camino_helpfile}
\title{Read Camino Command Help File}
\usage{
read_camino_helpfile(funcname)
}
\arguments{
\item{funcname}{Name of function to be called}
}
\value{
Character vector
}
\description{
Wrapper for finding the manual for the Camino function
and reading it into a text file
}
\examples{
read_camino_helpfile("voxel2scanner")
}
