% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shformatconverter.R
\name{shformatconverter}
\alias{shformatconverter}
\title{Wrapper for Camino \code{shformatconverter} function}
\usage{
shformatconverter()
}
\description{
Performs the Camino \code{shformatconverter} function
}

