% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camino_image2voxel.R, R/image2voxel.R
\name{camino_image2voxel}
\alias{camino_image2voxel}
\title{Wrapper for Camino \code{image2voxel} function}
\usage{
camino_image2voxel()

camino_image2voxel()
}
\arguments{
\item{infile}{Input 4D image}

\item{outfile}{Output filename for camino format}

\item{outputdatatype}{Output data type}

\item{verbose}{Print diagnostic messages}
}
\description{
Performs the Camino \code{image2voxel} function

Performs the Camino \code{image2voxel} function
}

